chore:#01 cameraがundefine
