+++ src/070_refactoring/start/index.html
@@ -3,7 +3,7 @@
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <script src="./index.js" type="module"></script>
+    <script src="./scripts/index.js" type="module"></script>
     <title>WebGL Practical</title>
   </head>
   <body>
