+++ src/070_refactoring/start/scripts/bootstrap.js
@@ -1,42 +1,39 @@
-import "./style.scss";
-
 import gsap from "gsap";
 import Scrollbar from "smooth-scrollbar";
 import { ScrollTrigger } from "gsap/ScrollTrigger";
-import {
-  WebGLRenderer,
-  Scene,
-  PerspectiveCamera,
-  PlaneGeometry,
-  MeshBasicMaterial,
-  ShaderMaterial,
-  Mesh,
-  Raycaster,
-  Vector2,
-} from "three";
+// import {
+//   WebGLRenderer,
+//   Scene,
+//   PerspectiveCamera,
+//   PlaneGeometry,
+//   MeshBasicMaterial,
+//   ShaderMaterial,
+//   Mesh,
+//   Raycaster,
+//   Vector2,
+// } from "three";
 
-const world = {};
-const os = [];
+// const world = {};
+// const os = [];
 const canvas = document.querySelector("#canvas");
 const canvasRect = canvas.getBoundingClientRect();
 
-const raycaster = new Raycaster();
-const pointer = new Vector2();
+// const raycaster = new Raycaster();
+// const pointer = new Vector2();
 
-init();
-function init() {
+export function init() {
   initScroller();
   bindResizeEvents();
 
-  world.renderer = new WebGLRenderer({
-    canvas,
-    antialias: true,
-  });
-  world.renderer.setSize(canvasRect.width, canvasRect.height, false);
-  world.renderer.setPixelRatio(window.devicePixelRatio);
-  world.renderer.setClearColor(0x000000, 0);
+  // world.renderer = new WebGLRenderer({
+  //   canvas,
+  //   antialias: true,
+  // });
+  // world.renderer.setSize(canvasRect.width, canvasRect.height, false);
+  // world.renderer.setPixelRatio(window.devicePixelRatio);
+  // world.renderer.setClearColor(0x000000, 0);
 
-  world.scene = new Scene();
+  // world.scene = new Scene();
 
   const cameraWidth = canvasRect.width;
   const cameraHeight = canvasRect.height;
@@ -46,75 +43,75 @@ function init() {
   const cameraZ = 2000;
   const radian = 2 * Math.atan(cameraHeight / 2 / cameraZ);
   const fov = radian * (180 / Math.PI);
-  world.camera = new PerspectiveCamera(fov, aspect, near, far);
-  world.camera.position.z = cameraZ;
+  // world.camera = new PerspectiveCamera(fov, aspect, near, far);
+  // world.camera.position.z = cameraZ;
 
-  const els = document.querySelectorAll("[data-webgl]");
-  els.forEach((el) => {
-    const rect = el.getBoundingClientRect();
+  // const els = document.querySelectorAll("[data-webgl]");
+  // els.forEach((el) => {
+  //   const rect = el.getBoundingClientRect();
 
-    const geometry = new PlaneGeometry(rect.width, rect.height, 1, 1);
-    // const material = new MeshBasicMaterial({
-    //   color: 0xff0000,
-    //   transparent: true,
-    //   opacity: 0.3,
-    // });
-    const material = new ShaderMaterial({
-      vertexShader: `
-        varying vec2 vUv;
+  //   const geometry = new PlaneGeometry(rect.width, rect.height, 1, 1);
+  //   // const material = new MeshBasicMaterial({
+  //   //   color: 0xff0000,
+  //   //   transparent: true,
+  //   //   opacity: 0.3,
+  //   // });
+  //   const material = new ShaderMaterial({
+  //     vertexShader: `
+  //       varying vec2 vUv;
 
-        void main() {
-          vUv = uv;
-          gl_Position = projectionMatrix * modelViewMatrix * vec4(position, 1.);
-        }
-      `,
-      fragmentShader: `
-        varying vec2 vUv;
-        uniform vec2 uMouse;
-        uniform float uHover;
+  //       void main() {
+  //         vUv = uv;
+  //         gl_Position = projectionMatrix * modelViewMatrix * vec4(position, 1.);
+  //       }
+  //     `,
+  //     fragmentShader: `
+  //       varying vec2 vUv;
+  //       uniform vec2 uMouse;
+  //       uniform float uHover;
 
-        void main() {
-          vec2 mouse = step(uMouse, vUv);
-          gl_FragColor = vec4(mouse, uHover, 1.);
-        }
-      `,
-      uniforms: {
-        uMouse: { value: new Vector2(0.5,0.5) },
-        uHover: { value: 0 }
-      }
-    })
-    const mesh = new Mesh(geometry, material);
-    mesh.position.z = 0;
+  //       void main() {
+  //         vec2 mouse = step(uMouse, vUv);
+  //         gl_FragColor = vec4(mouse, uHover, 1.);
+  //       }
+  //     `,
+  //     uniforms: {
+  //       uMouse: { value: new Vector2(0.5, 0.5) },
+  //       uHover: { value: 0 },
+  //     },
+  //   });
+  //   const mesh = new Mesh(geometry, material);
+  //   mesh.position.z = 0;
 
-    const { x, y } = getWorldPosition(rect, canvasRect);
-    mesh.position.x = x;
-    mesh.position.y = y;
+  //   const { x, y } = getWorldPosition(rect, canvasRect);
+  //   mesh.position.x = x;
+  //   mesh.position.y = y;
 
-    const o = {
-      mesh,
-      geometry,
-      material,
-      rect,
-      $: {
-        el,
-      },
-    };
+  //   const o = {
+  //     mesh,
+  //     geometry,
+  //     material,
+  //     rect,
+  //     $: {
+  //       el,
+  //     },
+  //   };
 
-    world.scene.add(mesh);
-    os.push(o);
-  });
+  //   world.scene.add(mesh);
+  //   os.push(o);
+  // });
 
-  render();
-  function render() {
-    requestAnimationFrame(render);
-    // スクロール処理
-    os.forEach((o) => scroll(o));
+  // render();
+  // function render() {
+  //   requestAnimationFrame(render);
+  //   // スクロール処理
+  //   os.forEach((o) => scroll(o));
 
-    // レイキャスティング
-    raycast();
+  //   // レイキャスティング
+  //   raycast();
 
-    world.renderer.render(world.scene, world.camera);
-  }
+  //   world.renderer.render(world.scene, world.camera);
+  // }
 }
 
 function scroll(o) {
@@ -223,50 +220,45 @@ function bindResizeEvents() {
   });
 }
 
+function onPointerMove(event) {
+  // calculate pointer position in normalized device coordinates
+  // (-1 to +1) for both components
 
-
-function onPointerMove( event ) {
-
-	// calculate pointer position in normalized device coordinates
-	// (-1 to +1) for both components
-
-	pointer.x = ( event.clientX / window.innerWidth ) * 2 - 1;
-	pointer.y = - ( event.clientY / window.innerHeight ) * 2 + 1;
-
+  pointer.x = (event.clientX / window.innerWidth) * 2 - 1;
+  pointer.y = -(event.clientY / window.innerHeight) * 2 + 1;
 }
 
-function raycast() {
+// function raycast() {
+//   // update the picking ray with the camera and pointer position
+//   raycaster.setFromCamera(pointer, world.camera);
 
-	// update the picking ray with the camera and pointer position
-	raycaster.setFromCamera( pointer, world.camera );
+//   // calculate objects intersecting the picking ray
+//   const intersects = raycaster.intersectObjects(world.scene.children);
+//   const intersect = intersects[0];
 
-	// calculate objects intersecting the picking ray
-	const intersects = raycaster.intersectObjects( world.scene.children );
-  const intersect = intersects[0];
-  
-	for ( let i = 0; i < world.scene.children.length; i ++ ) {
-    const _mesh = world.scene.children[i];
+//   for (let i = 0; i < world.scene.children.length; i++) {
+//     const _mesh = world.scene.children[i];
 
-    const uHover = _mesh.material.uniforms.uHover;
-    if(intersect?.object === _mesh) {
-      _mesh.material.uniforms.uMouse.value = intersect.uv;
-      uHover.__endValue = 1;
-    } else {
-      uHover.__endValue = 0;
-    }
+//     const uHover = _mesh.material.uniforms.uHover;
+//     if (intersect?.object === _mesh) {
+//       _mesh.material.uniforms.uMouse.value = intersect.uv;
+//       uHover.__endValue = 1;
+//     } else {
+//       uHover.__endValue = 0;
+//     }
 
-    uHover.value = lerp(uHover.value, uHover.__endValue, .1);
-	}
-}
+//     uHover.value = lerp(uHover.value, uHover.__endValue, 0.1);
+//   }
+// }
 
 // 線形補間
 function lerp(a, b, n) {
   let current = (1 - n) * a + n * b;
-  if(Math.abs(b - current) < 0.001) current = b;
+  if (Math.abs(b - current) < 0.001) current = b;
   return current;
 }
 
-window.addEventListener( 'pointermove', onPointerMove );
+window.addEventListener("pointermove", onPointerMove);
 
 /**
  * Raycasting
