+++ src/070_refactoring/start/scripts/glsl/world.js
@@ -1,0 +1,133 @@
+import { viewport } from "../helper/viewport";
+
+import {
+  WebGLRenderer,
+  Scene,
+  PerspectiveCamera,
+  PlaneGeometry,
+  MeshBasicMaterial,
+  ShaderMaterial,
+  Mesh,
+  Raycaster,
+  Vector2,
+} from "three";
+
+const world = {
+  init,
+  render,
+};
+
+const os = [];
+
+const raycaster = new Raycaster();
+const pointer = new Vector2();
+
+function init() {
+  world.renderer = new WebGLRenderer({
+    canvas,
+    antialias: true,
+  });
+  world.renderer.setSize(canvasRect.width, canvasRect.height, false);
+  world.renderer.setPixelRatio(window.devicePixelRatio);
+  world.renderer.setClearColor(0x000000, 0);
+
+  world.scene = new Scene();
+
+  world.camera = new PerspectiveCamera(
+    viewport.fov,
+    viewport.aspect,
+    viewport.near,
+    viewport.far
+  );
+  world.camera.position.z = cameraZ;
+
+  const els = document.querySelectorAll("[data-webgl]");
+  els.forEach((el) => {
+    const rect = el.getBoundingClientRect();
+
+    const geometry = new PlaneGeometry(rect.width, rect.height, 1, 1);
+    // const material = new MeshBasicMaterial({
+    //   color: 0xff0000,
+    //   transparent: true,
+    //   opacity: 0.3,
+    // });
+    const material = new ShaderMaterial({
+      vertexShader: `
+              varying vec2 vUv;
+      
+              void main() {
+                vUv = uv;
+                gl_Position = projectionMatrix * modelViewMatrix * vec4(position, 1.);
+              }
+            `,
+      fragmentShader: `
+              varying vec2 vUv;
+              uniform vec2 uMouse;
+              uniform float uHover;
+      
+              void main() {
+                vec2 mouse = step(uMouse, vUv);
+                gl_FragColor = vec4(mouse, uHover, 1.);
+              }
+            `,
+      uniforms: {
+        uMouse: { value: new Vector2(0.5, 0.5) },
+        uHover: { value: 0 },
+      },
+    });
+    const mesh = new Mesh(geometry, material);
+    mesh.position.z = 0;
+
+    const { x, y } = getWorldPosition(rect, canvasRect);
+    mesh.position.x = x;
+    mesh.position.y = y;
+
+    const o = {
+      mesh,
+      geometry,
+      material,
+      rect,
+      $: {
+        el,
+      },
+    };
+
+    world.scene.add(mesh);
+    os.push(o);
+  });
+}
+function render() {
+  requestAnimationFrame(render);
+  // スクロール処理
+  os.forEach((o) => scroll(o));
+
+  // レイキャスティング
+  raycast();
+
+  world.renderer.render(world.scene, world.camera);
+}
+
+function raycast() {
+  // update the picking ray with the camera and pointer position
+  raycaster.setFromCamera(pointer, world.camera);
+
+  // calculate objects intersecting the picking ray
+  const intersects = raycaster.intersectObjects(world.scene.children);
+  const intersect = intersects[0];
+
+  for (let i = 0; i < world.scene.children.length; i++) {
+    const _mesh = world.scene.children[i];
+
+    const uHover = _mesh.material.uniforms.uHover;
+    if (intersect?.object === _mesh) {
+      _mesh.material.uniforms.uMouse.value = intersect.uv;
+      uHover.__endValue = 1;
+    } else {
+      uHover.__endValue = 0;
+    }
+
+    uHover.value = lerp(uHover.value, uHover.__endValue, 0.1);
+  }
+}
+
+export default world;
